feat: added discord icon
